+++ Sources/SwiftUIKit/_Deprecated/Presentation/AlertContext.swift
@@ -33,6 +33,7 @@ import SwiftUI
 /// This view modifier will also inject the provided context
 /// as an environment object into the view hierarchy, to let
 /// other views in the same hierarchy reuse the same context.
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
 public class AlertContext: PresentationContext<Alert> {
 
     @MainActor
@@ -45,14 +46,46 @@ public class AlertContext: PresentationContext<Alert> {
 
 public extension View {
     
-    /// Bind an ``AlertContext`` to the view.
-    ///
-    /// This also injects this context as environment object.
+    @available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
     func alert(_ context: AlertContext) -> some View {
         alert(
             isPresented: context.isActiveBinding,
             content: context.content ?? { Alert(title: Text("")) }
         )
         .environmentObject(context)
+    }
+}
+
+/**
+ This key can be used to keep track of the currently focused
+ context within a multi-window app.
+
+ To apply it to a view, just use the `focusedValue` modifier
+ with your context instance:
+
+ ```swift
+ .focusedValue(\.fullscreenCoverContext, context)
+ ```
+
+ To access the focus value, e.g. in a menu command, just use
+ the `@FocusedValue` property wrapper:
+
+ ```swift
+ @FocusedValue(\.fullscreenCoverContext)
+ var fullscreenCoverContext: FullscreenCoverContext?
+ ```
+ */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
+public struct AlertContextFocusedValueKey: FocusedValueKey {
+
+    public typealias Value = AlertContext
+}
+
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
+public extension FocusedValues {
+
+    var alertContext: AlertContextFocusedValueKey.Value? {
+        get { self[AlertContextFocusedValueKey.self] }
+        set { self[AlertContextFocusedValueKey.self] = newValue }
     }
 }