+++ Sources/SwiftUIKit/_Deprecated/Presentation/SheetContext.swift
@@ -32,6 +32,7 @@ import SwiftUI
  an environment object into the view hierarchy, to let other
  views in the same view hierarchy reuse the same context.
  */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
 public class SheetContext: PresentationContext<AnyView> {
 
     @MainActor
@@ -40,6 +41,7 @@ public class SheetContext: PresentationContext<AnyView> {
     }
 }
 
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
 public extension View {
     
     /// Bind an ``SheetContext`` to the view.
@@ -51,5 +53,39 @@ public extension View {
             content: context.content ?? { AnyView(EmptyView()) }
         )
         .environmentObject(context)
+    }
+}
+
+/**
+ This key can be used to keep track of the currently focused
+ context within a multi-window app.
+
+ To apply it to a view, just use the `focusedValue` modifier
+ with your context instance:
+
+ ```swift
+ .focusedValue(\.sheetContext, context)
+ ```
+
+ To access the focus value, e.g. in a menu command, just use
+ the `@FocusedValue` property wrapper:
+
+ ```swift
+ @FocusedValue(\.sheetContext)
+ private var sheetContext: SheetContext?
+ ```
+ */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
+public struct SheetContextFocusedValueKey: FocusedValueKey {
+
+    public typealias Value = SheetContext
+}
+
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
+public extension FocusedValues {
+
+    var sheetContext: SheetContextFocusedValueKey.Value? {
+        get { self[SheetContextFocusedValueKey.self] }
+        set { self[SheetContextFocusedValueKey.self] = newValue }
     }
 }