+++ RELEASE_NOTES.md
@@ -12,7 +12,11 @@ SwiftUIKit makes its best effort to honor semver, but breaking changes can occur
 
 ### 🗑️ Deprecations
 
+* `AlertContext` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `PresentationContext` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `FullScreenCoverContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `ScanCodeType` has been moved to https://github.com/danielsaidi/ScanCodes.
+* `SheetContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `Image+ScanCode` has been moved to https://github.com/danielsaidi/ScanCodes.